lt_crm/requirements.txt
flask==3.0.0
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.2
flask-babel==3.1.0
flask-talisman==1.1.0
celery==5.3.4
redis==5.0.1
marshmallow==3.20.1
flask-restx==1.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
PyJWT==2.8.0
flask-limiter==3.5.0
pandas==2.1.1
openpyxl==3.1.2
reportlab==4.0.7
flask-minify==0.42
webassets==2.0
babel==2.13.1
httpx==0.25.1
prometheus-flask-exporter==0.22.4
sentry-sdk==1.34.0
Pillow==10.1.0
requests==2.31.0 